move handle_hook to avoid circle import, add lookup resolve to handle_hooks
